AI Controller added (please read the description)
This AI Controller is to call RunBehaviorTree(btree); method.

This method is only callable under Controller class scope and its extensions. Since all the movement logics are implemented in Character class, there might be a need of migrating all the codes under this Controller.

I assume that the tutor did not implemented the movement implementations in Controller code since the class hierarchy is too complicated in that way. (Character is still a child class of Pawn therefore it can call the simple movement things.)