Bug Fixed
after submission I noticed this bug where it can no longer compile or buid. function declaration of unused function was revived somehow, and I did not noticed it.

I asked in the comment to regard the substitutional file uploaded in Google drive due to this reason. hopefully the tutor will consider.